move WeightDecay to regularier test=develop
